Backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-groq
langchain-cerebras
langgraph
langchain-tavily
pymupdf
opencv-python
pytesseract
assemblyai
yt-dlp
numpy
pydantic
python-multipart